Separate data size validation of each unique record
 - include refactor of message formatter, change datatype of 2nd argument to support additional messages

+++ metering-app/src/main/java/com/pemc/crss/metering/parser/bcq/BcqInterval.java
@@ -40,7 +40,7 @@ public enum BcqInterval {
         return timeInMillis;
     }
 
-    public long getValidNoOfRecords() {
+    public int getValidNoOfRecords() {
         return validNoOfRecords;
     }
